refactor: extract condition evaluator from context.rs
Split context.rs (1074 LOC) into:
- context.rs (307 LOC): RecipeContext, template rendering, delegates evaluate()
- condition.rs (765 LOC): tokenizer, recursive-descent parser, built-in functions/methods

RecipeContext::evaluate() now delegates to condition::evaluate_condition().
All types re-exported: ConditionError remains accessible via crate::condition.

No behavioral changes. All 306 tests pass.
